requirements_new.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
anyio==3.7.1
async-timeout==5.0.1
attrs==25.3.0
bcrypt==4.3.0
cachetools==5.5.2
cryptography==44.0.2
cssselect2==0.8.0
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.1
et_xmlfile==2.0.0
fastapi==0.103.1
fonttools==4.57.0
frozenlist==1.6.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.167.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.72.0rc1
grpcio-status==1.71.0
gspread==6.2.0
httpcore==1.0.8
httplib2==0.22.0
httpx==0.28.1
jiter==0.9.0
lxml==5.4.0
multidict==6.4.3
oauth2client==4.1.3
oauthlib==3.2.2
openai==0.28.0
openpyxl==3.1.5
packaging==25.0
pandas==1.4.2
passlib==1.7.4
pcapy==0.11.4
pdf2image==1.17.0
pdfminer.six==20250327
pdfplumber==0.11.6
pillow==11.2.1
prettytable==3.15.1
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.4.8
pyasn1_modules==0.4.2
pydyf==0.11.0
pyexiv2==2.15.3
pymongo==4.12.1
pyparsing==3.2.3
pypdfium2==4.30.1
pyphen==0.17.2
pytesseract==0.3.13
python-docx==1.1.2
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
python-nmap==0.7.1
pyzmail36==1.0.5
requests-oauthlib==2.0.0
rsa==4.9.1
scapy==2.6.1
schedule==1.2.2
starlette==0.27.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tqdm==4.67.1
twitter==1.19.6
uritemplate==4.1.1
wcwidth==0.2.13
weasyprint==65.1
webencodings==0.5.1
yarl==1.20.0
zopfli==0.2.3.post1

